Made add_background_rectangle are more general method for all mobjects
